do code refactoring
