+++ src/components/SingleMessage/SingleMessage.js
@@ -2,25 +2,14 @@ import React from 'react'
 import './SingleMessage.css';
 
 const SingleMessage = ({ left , text }) => {
-  if (left) {
     return (
       <div className="chatLeftContainer">
         <div className="user">You</div>
-          <div className='chat-left'>
+          <div className= { left ? 'chat-left' : 'chat-right'}>
             {text}
           </div>
       </div>
     )
-  } else {
-    return (
-      <div className="chatLeftContainer">
-        <div className="user">Assistant</div>
-          <div className='chat-right'>
-            {text}
-        </div>
-      </div>
-    )
-  }
 }
 
 export default SingleMessage;
